rename error.ts to errors.ts
